将Revision 从 int 改为 long。移除 NiumaInventory.Enum.ItemType / ItemQuality 依赖改为 ItemTypeKey / QualityKey后续 ShopService 负责把背包物品定义映射成表现 Key，UI 不需要直接依赖 Inventory 枚举+初代服务层

+++ Runtime/Data/ShopProgressSnapshot.cs
@@ -24,7 +24,7 @@ namespace NiumaShop.Data
         /// 单商店修订号。
         /// 库存、限购、解锁、折扣、刷新状态变化时递增。
         /// </summary>
-        public int Revision;
+        public long Revision;
 
         /// <summary>
         /// 商品运行时快照。
